pyproject.toml
[project]
name = "gettasker"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "aiofiles (==24.1.0)",
    "alembic (==1.14.1)",
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "asyncpg (==0.30.0)",
    "bcrypt (==4.2.1)",
    "black (==25.1.0)",
    "certifi (==2025.1.31)",
    "cffi (==1.17.1)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "cryptography (==44.0.1)",
    "dnspython (==2.7.0)",
    "ecdsa (==0.19.0)",
    "email-validator (==2.2.0)",
    "fastapi (==0.115.8)",
    "fastapi-cache2 (==0.2.2)",
    "fastapi-cli (==0.0.7)",
    "greenlet (==3.1.1)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.5)",
    "mako (==1.3.9)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "mdurl (==0.1.2)",
    "mypy-extensions (==1.0.0)",
    "orjson (==3.10.15)",
    "packaging (==24.2)",
    "passlib (==1.7.4)",
    "pathspec (==0.12.1)",
    "pendulum (==3.0.0)",
    "platformdirs (==4.3.6)",
    "pyasn1 (==0.4.8)",
    "pycparser (==2.22)",
    "pydantic (==2.10.6)",
    "pydantic-extra-types (==2.10.2)",
    "pydantic-settings (==2.7.1)",
    "pydantic-core (==2.27.2)",
    "pygments (==2.19.1)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.0.1)",
    "python-jose (==3.4.0)",
    "python-multipart (==0.0.20)",
    "pyyaml (==6.0.2)",
    "redis (==5.0.8)",
    "rich (==13.9.4)",
    "rich-toolkit (==0.13.2)",
    "rsa (==4.9)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.38)",
    "starlette (==0.45.3)",
    "time-machine (==2.16.0)",
    "typer (==0.15.1)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2025.1)",
    "ujson (==5.10.0)",
    "uvicorn (==0.34.0)",
    "watchfiles (==1.0.4)",
    "websockets (==14.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.1.1"
pytest-mock = "^3.14.1"

